+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Enabled `positionMode: natural` for tooltips by default, can be changed via tooltip options
 - Exposed type checking helpers `isArray()`, `isSet()` and `isRegExp()`
 - Fixed `badge` view options when it receives an array as data
+- Fixed `struct` view annotation styling to avoid treating `href` values as links when they are falsy
 - Fixed a sticky tooltip for "Enable view inspection" button on inspecting start
 - Fixed the delayed popup display after calling the `hide()` method
 - Fixed positioning of popups with `position: pointer`